SettingContents 파일 분리
